checklist id changed from Integer to Long

+++ global-service/src/main/java/in/cdac/university/globalService/bean/CheckListBean.java
@@ -31,7 +31,7 @@ public class CheckListBean {
 
 	private Long unumNdtlId;
 
-	private Integer unumChecklistId;
+	private Long unumChecklistId;
 
 	private Long unumMtempleId;
